requirements/dev.txt
# Additional requirements useful for development.

-r base.txt

mock==0.7.2
coverage==3.4

Fabric==1.0.1
pycrypto==2.3
paramiko==1.7.7.1

ipython==0.10.2

django-webtest==1.3
WebTest==1.2.3
WebOb==1.0.7
